fix: db crashes when used with a connection pool
The database we use is behind a pgbouncer setup. PGbouncer doesn’t support statement_timeout and when we issue queries with it the queries are failing with:
Unhandled rejection: error: unsupported startup parameter: statement_timeout

If one uses a connection pool/pgbouncer it’s not necessary to use statement_timeout as the pool takes care of the connection.